requirements.txt
fastapi==0.135.2
uvicorn[standard]>=0.30
openai==2.30.0
pydantic==2.12.5
python-dotenv>=1.0.0
sqlalchemy>=2.0
bcrypt>=4.0
PyJWT>=2.8
email-validator>=2.0
slowapi>=0.1.9
psycopg2-binary>=2.9
